requirements.txt
groq>=1.0.0
pydantic
python-dotenv